feat(prometheus): Restructure prometheus_metrics package
